move bad encoding fraud proof to share pkg (upd files)
