+++ share/eds/retriever_test.go
@@ -2,18 +2,22 @@ package eds
 
 import (
 	"context"
+	"errors"
 	"math/rand"
 	"testing"
 	"time"
 
+	"github.com/ipfs/go-blockservice"
 	mdutils "github.com/ipfs/go-merkledag/test"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/celestiaorg/celestia-app/pkg/da"
 	"github.com/celestiaorg/celestia-app/pkg/wrapper"
-
+	"github.com/celestiaorg/celestia-node/fraud"
+	"github.com/celestiaorg/celestia-node/header"
 	"github.com/celestiaorg/celestia-node/share"
+	"github.com/celestiaorg/celestia-node/share/eds/byzantine"
 	"github.com/celestiaorg/celestia-node/share/ipld"
 	"github.com/celestiaorg/nmt"
 	"github.com/celestiaorg/rsmt2d"
@@ -90,7 +94,7 @@ func TestRetriever_ByzantineError(t *testing.T) {
 	dah := da.NewDataAvailabilityHeader(attackerEDS)
 	r := NewRetriever(bserv)
 	_, err = r.Retrieve(ctx, &dah)
-	var errByz *share.ErrByzantine
+	var errByz *byzantine.ErrByzantine
 	require.ErrorAs(t, err, &errByz)
 }
 
@@ -122,4 +126,32 @@ func TestRetriever_MultipleRandQuadrants(t *testing.T) {
 
 	_, err = ses.Reconstruct(ctx)
 	assert.NoError(t, err)
+}
+
+func TestFraudProofValidation(t *testing.T) {
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second*15)
+	defer t.Cleanup(cancel)
+	bServ := mdutils.Bserv()
+
+	var errByz *byzantine.ErrByzantine
+	faultHeader, err := generateByzantineError(ctx, t, bServ)
+	require.True(t, errors.As(err, &errByz))
+
+	p := byzantine.CreateBadEncodingProof([]byte("hash"), uint64(faultHeader.Height), errByz)
+	err = p.Validate(faultHeader)
+	require.NoError(t, err)
+}
+
+func generateByzantineError(
+	ctx context.Context,
+	t *testing.T,
+	bServ blockservice.BlockService,
+) (*header.ExtendedHeader, error) {
+	_, store := fraud.CreateTestService(t, false)
+	h, err := store.GetByHeight(ctx, 1)
+	require.NoError(t, err)
+
+	faultHeader := header.CreateFraudExtHeader(t, h, bServ)
+	_, err = NewRetriever(bServ).Retrieve(ctx, faultHeader.DAH)
+	return faultHeader, err
 }
